feat: Add description field to Contact model and update related views
This commit adds a new `descriptionContact` field to the `Contact` model, allowing for the storage of a description for each contact. The `ContactController` has been updated to handle the new field in the `store` and `update` methods. Additionally, the `edit.blade.php` and `index.blade.php` views have been modified to include a textarea input for the `descriptionContact` field.